Function linearkalman::filter_step

pub fn filter_step(kalman_filter: &KalmanFilter,
                   init: &KalmanState,
                   measure: &Vector<f64>)
                   -> (KalmanState, KalmanState)

Returns a tuple containing updated and predicted estimates (in that order) of the state variable and its covariance. This function might be useful for cases where data is incoming and being updated in real-time so that Kalman filtering is run incrementally. Note that given some initial values for x and P, filter_step makes a prediction and then runs the update step to correct the prediction based on the observed data.