Function linearkalman::update_step

pub fn update_step(kalman_filter: &KalmanFilter,
                   pred: &KalmanState,
                   measure: &Vector<f64>)
                   -> KalmanState

Returns an updated state variable mean and covariance given predicted and observed data. Typically, update step will be called after prediction step, data of which will be consequently used as input in updating.